requirements-core.txt
# Core Requirements - Only what you actually use
# Install: pip install -r requirements-core.txt

# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Statistics
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# NLP - Basic
nltk>=3.8.0

# Utilities
tqdm>=4.65.0
jupyter>=1.0.0
ipywidgets>=8.0.0
python-dotenv>=1.0.0

# OpenAI (for GPT summarization)
openai>=1.0.0
